spell_checker
spell_checker is a way to see if a word exists in a list of words using a bloom filter to do so. Note, not all words may be contained in this list, but "mywords.txt" is my text document that contains a list of around 400000 words that I compiled together.